refactor xliff commandline
